Move runtime-independent CUDA functionality to async_cuda_base module
