+++ libs/pika/async_cuda/CMakeLists.txt
@@ -10,38 +10,17 @@ endif()
 
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
 
-# Default location is $PIKA_ROOT/libs/async_cuda/include
 set(async_cuda_headers
-    pika/async_cuda/cublas_exception.hpp
-    pika/async_cuda/cublas_handle.hpp
-    pika/async_cuda/cuda_device_scope.hpp
-    pika/async_cuda/cuda_event.hpp
     pika/async_cuda/cuda_polling_helper.hpp
     pika/async_cuda/cuda_pool.hpp
     pika/async_cuda/cuda_scheduler.hpp
     pika/async_cuda/cuda_scheduler_bulk.hpp
-    pika/async_cuda/cuda_stream.hpp
-    pika/async_cuda/cusolver_exception.hpp
-    pika/async_cuda/cusolver_handle.hpp
-    pika/async_cuda/custom_blas_api.hpp
-    pika/async_cuda/custom_lapack_api.hpp
-    pika/async_cuda/detail/cuda_debug.hpp
     pika/async_cuda/detail/cuda_event_callback.hpp
     pika/async_cuda/then_on_host.hpp
     pika/async_cuda/then_with_stream.hpp
 )
 
-set(async_cuda_sources
-    cublas_exception.cpp
-    cublas_handle.cpp
-    cuda_device_scope.cpp
-    cuda_event_callback.cpp
-    cuda_pool.cpp
-    cuda_scheduler.cpp
-    cuda_stream.cpp
-    cusolver_exception.cpp
-    cusolver_handle.cpp
-)
+set(async_cuda_sources cuda_event_callback.cpp cuda_pool.cpp cuda_scheduler.cpp)
 
 if(PIKA_WITH_HIP)
   set(async_cuda_extra_deps roc::rocblas roc::rocsolver)
@@ -58,17 +37,18 @@ pika_add_module(
   SOURCES ${async_cuda_sources}
   HEADERS ${async_cuda_headers}
   MODULE_DEPENDENCIES
-    pika_allocator_support
     pika_assertion
     pika_async_base
+    pika_async_cuda_base
     pika_concurrency
     pika_config
-    pika_debugging
+    pika_coroutines
     pika_errors
+    pika_execution
     pika_execution_base
-    pika_memory
     pika_runtime
     pika_threading_base
+    pika_topology
   DEPENDENCIES ${async_cuda_extra_deps}
   CMAKE_SUBDIRS examples tests
 )
