+++ libs/pika/async_cuda/include/pika/async_cuda/cuda_scheduler_bulk.hpp
@@ -8,8 +8,8 @@
 
 #include <pika/assert.hpp>
 #include <pika/async_cuda/cuda_pool.hpp>
-#include <pika/async_cuda/cuda_stream.hpp>
 #include <pika/async_cuda/then_with_stream.hpp>
+#include <pika/async_cuda_base/cuda_stream.hpp>
 #include <pika/concepts/concepts.hpp>
 #include <pika/execution/algorithms/bulk.hpp>
 #include <pika/execution/algorithms/execute.hpp>
