+++ libs/pika/async_cuda/src/cuda_event_callback.cpp
@@ -9,9 +9,9 @@
 
 #include <pika/config.hpp>
 #include <pika/assert.hpp>
-#include <pika/async_cuda/cuda_event.hpp>
-#include <pika/async_cuda/detail/cuda_debug.hpp>
 #include <pika/async_cuda/detail/cuda_event_callback.hpp>
+#include <pika/async_cuda_base/cuda_event.hpp>
+#include <pika/async_cuda_base/detail/cuda_debug.hpp>
 #include <pika/concurrency/concurrentqueue.hpp>
 #include <pika/concurrency/spinlock.hpp>
 #include <pika/datastructures/detail/small_vector.hpp>
