+++ libs/pika/async_cuda_base/README.rst
@@ -1,0 +1,13 @@
+
+..
+    Copyright (c) 2019 The STE||AR-Group
+
+    SPDX-License-Identifier: BSL-1.0
+    Distributed under the Boost Software License, Version 1.0. (See accompanying
+    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+===============
+async_cuda_base
+===============
+
+This library is part of pika.
