+++ libs/pika/async_cuda_base/docs/index.rst
@@ -1,0 +1,14 @@
+..
+    Copyright (c) 2019 The STE||AR-Group
+
+    SPDX-License-Identifier: BSL-1.0
+    Distributed under the Boost Software License, Version 1.0. (See accompanying
+    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+.. _modules_async_cuda_base:
+
+===============
+async_cuda_base
+===============
+
+This module defines basic CUDA and HIP utilities that are independent of the pika runtime.
