+++ libs/pika/async_cuda_base/include/pika/async_cuda_base/cublas_handle.hpp
@@ -7,8 +7,8 @@
 #pragma once
 
 #include <pika/config.hpp>
-#include <pika/async_cuda/cuda_stream.hpp>
-#include <pika/async_cuda/custom_blas_api.hpp>
+#include <pika/async_cuda_base/cuda_stream.hpp>
+#include <pika/async_cuda_base/custom_blas_api.hpp>
 
 #include <fmt/format.h>
 #include <whip.hpp>
