+++ libs/pika/async_cuda_base/include/pika/async_cuda_base/cusolver_exception.hpp
@@ -8,7 +8,7 @@
 
 #include <pika/config.hpp>
 #if defined(PIKA_HAVE_GPU_SUPPORT)
-# include <pika/async_cuda/custom_lapack_api.hpp>
+# include <pika/async_cuda_base/custom_lapack_api.hpp>
 # include <pika/errors/exception.hpp>
 
 # include <string>
