+++ libs/pika/async_cuda_base/include/pika/async_cuda_base/cusolver_handle.hpp
@@ -8,8 +8,8 @@
 
 #include <pika/config.hpp>
 #if defined(PIKA_HAVE_GPU_SUPPORT)
-# include <pika/async_cuda/cuda_stream.hpp>
-# include <pika/async_cuda/custom_lapack_api.hpp>
+# include <pika/async_cuda_base/cuda_stream.hpp>
+# include <pika/async_cuda_base/custom_lapack_api.hpp>
 
 # include <fmt/format.h>
 # include <whip.hpp>
