+++ libs/pika/async_cuda_base/src/cuda_device_scope.cpp
@@ -5,7 +5,7 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <pika/config.hpp>
-#include <pika/async_cuda/cuda_device_scope.hpp>
+#include <pika/async_cuda_base/cuda_device_scope.hpp>
 
 #include <whip.hpp>
