+++ libs/pika/async_cuda_base/tests/performance/CMakeLists.txt
@@ -1,0 +1,5 @@
+# Copyright (c) 2020 John Biddiscombe
+#
+# SPDX-License-Identifier: BSL-1.0
+# Distributed under the Boost Software License, Version 1.0. (See accompanying
+# file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
