+++ libs/pika/async_cuda_base/tests/unit/CMakeLists.txt
@@ -4,26 +4,11 @@
 # Distributed under the Boost Software License, Version 1.0. (See accompanying
 # file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-set(tests
-    cublas_matmul
-    cuda_bulk
-    cuda_device_reset
-    cuda_pool
-    cuda_scheduler
-    cuda_sender
-    then_with_stream
-)
+set(tests cublas_handle cuda_stream cusolver_handle)
 
-set(cublas_matmul_PARAMETERS THREADS 4)
-set(cuda_sender_PARAMETERS THREADS 4)
-set(then_with_stream_PARAMETERS THREADS 4)
-set(then_with_stream_DEPENDENCIES pika_execution_test_utilities)
-
-set(cuda_bulk_GPU ON)
-set(cuda_pool_GPU ON)
-set(cuda_scheduler_GPU ON)
-set(cuda_sender_GPU ON)
-set(then_with_stream_GPU ON)
+set(cublas_handle_GPU ON)
+set(cuda_stream_GPU ON)
+set(cusolver_handle_GPU ON)
 
 foreach(test ${tests})
   if(${${test}_GPU})
@@ -40,10 +25,10 @@ foreach(test ${tests})
     SOURCES ${sources} ${${test}_FLAGS}
     DEPENDENCIES ${${test}_DEPENDENCIES}
     EXCLUDE_FROM_ALL
-    FOLDER "Tests/Unit/Modules/AsyncCuda"
+    FOLDER "Tests/Unit/Modules/AsyncCudaBase"
   )
 
   pika_add_unit_test(
-    "modules.async_cuda" ${test} ${${test}_PARAMETERS} RUN_SERIAL
+    "modules.async_cuda_base" ${test} ${${test}_PARAMETERS} RUN_SERIAL
   )
 endforeach()
