+++ libs/pika/include/include/pika/cuda.hpp
@@ -9,4 +9,5 @@
 #include <pika/config.hpp>
 #if defined(PIKA_HAVE_GPU_SUPPORT)
 # include <pika/modules/async_cuda.hpp>
+# include <pika/modules/async_cuda_base.hpp>
 #endif
